Layout1dBase: Rename `layout-1d-base.js` to `Layout1dBase.js`; don't use default export.
