+++ src/layouts/Layout1dBase.js
@@ -1,4 +1,4 @@
-export default class Layout extends EventTarget {
+export class Layout1dBase extends EventTarget {
   constructor(config) {
     super();
 
@@ -316,4 +316,4 @@ export default class Layout extends EventTarget {
       [this._secondarySizeDim]: this._itemDim2,
     };
   }
-}
+};
