+++ src/layouts/layout-1d.js
@@ -1,4 +1,4 @@
-import Layout1dBase from './layout-1d-base.js';
+import {Layout1dBase} from './Layout1dBase.js';
 
 export default class Layout extends Layout1dBase {
   constructor(config) {
